+++ src/views/pages/civ5/civ5.vue
@@ -1,12 +1,12 @@
 <template>
-  <div class="page-civ-2 coa-padding-s">
+  <div class="page-civ-5 coa-padding-s">
     <h1 class="md-display-1">{{ pageTitle }}</h1>
     <!-- [ add here specific Civ-version content ] -->
   </div>
 </template>
 
 <script>
-const page = { version: 2, name: "civ2", label: "CIV-2" };
+const page = { version: 5, name: "civ5", label: "CIV-5" };
 
 export default {
   name: page.name,
@@ -33,6 +33,7 @@ export default {
   mounted() {
     console.log("mounted(): component did mount");
     this.myLabel = "MyLabel is now updated!!!";
+    console.groupEnd();
   },
 
   // Updating (Re-render)
@@ -67,7 +68,7 @@ export default {
 <style lang="scss" scoped>
 @import "@/styles/index.scss";
 
-.page-civ-2 {
+.page-civ-5 {
   margin: 0;
   padding: $size-s $size-m;
   background-color: $color-white;
